addUserIdForCreate And deleteAndEdit Unlink update

+++ controllers/note.js
@@ -51,6 +51,7 @@ exports.createNote = (req, res, next) => {
     title,
     content,
     cover_image: cover_image ? cover_image.path : "",
+    author: req.userId,
   })
     .then((_) => {
       return res.status(201).json({ message: "Note Created" });
@@ -64,6 +65,7 @@ exports.createNote = (req, res, next) => {
 exports.getNoteDetails = (req, res, next) => {
   const { id } = req.params;
   Note.findById(id)
+    .populate("author", "username")
     .then((note) => {
       return res.status(200).json(note);
     })
@@ -77,7 +79,9 @@ exports.deleteNote = (req, res, next) => {
   const { id } = req.params;
   Note.findById(id)
     .then((note) => {
-      unlink(note.cover_image);
+      if (note.cover_image) {
+        unlink(note.cover_image);
+      }
 
       return Note.findByIdAndDelete(id).then((_) => {
         return res.status(204).json({ message: "Note Deleted" });
@@ -114,13 +118,9 @@ exports.updateNote = (req, res, next) => {
       note.title = title;
       note.content = content;
       if (cover_image) {
-        unlink(note.cover_image, (err) => {
-          if (err) {
-            console.log(err);
-            return;
-          }
-          //file removed
-        });
+        if (note.cover_image) {
+          unlink(note.cover_image);
+        }
         note.cover_image = cover_image.path;
       }
       return note.save();
